move getMessageTextContent to core
